+++ packages/chrome-devtools/src/utils/chrome/override-remote.ts
@@ -1,6 +1,13 @@
-const basicProxyCore = require('../../vendor/basic-proxy-core.js') as {
-  registerOverridePlugin(globalObject?: typeof globalThis): unknown;
-};
+import {
+  BasicProxyCore,
+  resolveBasicProxyCore,
+} from './resolve-basic-proxy-core';
+
+const basicProxyCore = resolveBasicProxyCore(
+  require('../../vendor/basic-proxy-core.js') as
+    | BasicProxyCore
+    | { default: BasicProxyCore },
+);
 
 basicProxyCore.registerOverridePlugin(globalThis);