+++ packages/observability-plugin/src/runtime/preload.ts
@@ -1,0 +1,5 @@
+import type { GeneratePreloadAssetsResult } from '../type';
+
+export function continuePreloadAssetsGeneration(): GeneratePreloadAssetsResult {
+  return undefined as unknown as GeneratePreloadAssetsResult;
+}