+++ packages/metro-plugin-rnef/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @module-federation/metro-plugin-rnef
 
+## 2.9.0
+
+### Patch Changes
+
+- @module-federation/metro@2.9.0
+
+## 2.8.2
+
+### Patch Changes
+
+- @module-federation/metro@2.8.2
+
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies [1ed404a]
+  - @module-federation/metro@2.8.1
+
 ## 2.8.0
 
 ### Patch Changes