+++ packages/treeshake-frontend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/treeshake-frontend
 
+## 2.9.0
+
+## 2.8.2
+
+## 2.8.1
+
 ## 2.8.0
 
 ## 2.7.0